requirements.txt
pytest==8.2.0
selenium==4.9.1
ipython
google-generativeai
python-dotenv
notebook